feat(global_dependencies_provider): refactor widget to provide dependencies unsing `getIt` package
