+++ lib/app/view/app.dart
@@ -12,40 +12,37 @@ class App extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return RepositoriesProvider(
-      child: BlocsProvider(
-        child: Builder(
-          builder: (context) {
-            final authenticationCubit = context.read<AuthenticationCubit>();
-            final appRouter = AppRouter(
-              authenticationCubit: authenticationCubit,
-            );
-            return BlocSelector<SettingsBloc, SettingsState, AppBrightness>(
-              selector: (state) => state.brightness,
-              builder: (context, brightness) {
-                // make sure you don't initiate your router
-                // inside of the build function.
-                return MaterialApp.router(
-                  theme: ThemeData(
-                    useMaterial3: true,
-                    colorScheme: lightColorScheme,
-                    textTheme: textTheme,
-                  ),
-                  themeMode: brightness.themeMode,
-                  darkTheme: ThemeData(
-                    useMaterial3: true,
-                    colorScheme: darkColorScheme,
-                    textTheme: textTheme,
-                  ),
-                  localizationsDelegates:
-                      AppLocalizations.localizationsDelegates,
-                  supportedLocales: AppLocalizations.supportedLocales,
-                  routerConfig: appRouter.config(),
-                );
-              },
-            );
-          },
-        ),
+    return GlobalDependenciesProvider(
+      child: Builder(
+        builder: (context) {
+          final authenticationCubit = context.read<AuthenticationCubit>();
+          final appRouter = AppRouter(
+            authenticationCubit: authenticationCubit,
+          );
+          return BlocSelector<SettingsBloc, SettingsState, AppBrightness>(
+            selector: (state) => state.brightness,
+            builder: (context, brightness) {
+              // make sure you don't initiate your router
+              // inside of the build function.
+              return MaterialApp.router(
+                theme: ThemeData(
+                  useMaterial3: true,
+                  colorScheme: lightColorScheme,
+                  textTheme: textTheme,
+                ),
+                themeMode: brightness.themeMode,
+                darkTheme: ThemeData(
+                  useMaterial3: true,
+                  colorScheme: darkColorScheme,
+                  textTheme: textTheme,
+                ),
+                localizationsDelegates: AppLocalizations.localizationsDelegates,
+                supportedLocales: AppLocalizations.supportedLocales,
+                routerConfig: appRouter.config(),
+              );
+            },
+          );
+        },
       ),
     );
   }
